Sorting app container with proper flex settings
