+++ src/components/nav/Narbar.tsx
@@ -13,7 +13,7 @@ import {
     faTools,
     faUsers
 } from "@fortawesome/free-solid-svg-icons";
-import NavItem from "./NavItemProps";
+import NavItemLink from "./NavItemLink";
 
 const NavContainer = styled.nav<{$active?: boolean; }>`
   display: flex;
@@ -21,7 +21,7 @@ const NavContainer = styled.nav<{$active?: boolean; }>`
   
   background-color: ${Colours.DarkPrimary};
 
-  flex-shrink: 0;
+  //flex-shrink: 0;
   //width: 15vw;
   width: 100%;
   min-width: 15vw;
@@ -41,18 +41,12 @@ const NavItems = styled.ol`
   flex-direction: column;
   justify-content: space-evenly;
   gap: 8px;
-  //gap-row: 8px;
-  //width: 100%;
   width: 70%;
-  
-  margin: 0; //todo remove once global style
-  padding: 0;
 `;
 
 const NavItemOld = styled.li`
   padding: 1em;
   background-color: ${Colours.LightPrimary};
-  
   //width: 100%;
   
   border: #333333;
@@ -60,33 +54,23 @@ const NavItemOld = styled.li`
   //text-align: center;
 `;
 
-
-const NavText = styled.text`
-  color: ${Colours.White};
-  text-decoration: none;
-  
-  //text-decoration-style: ;
-`;
-
-
 const NarBar = () => {
-
 
     return (<NavContainer>
         <button>X</button>
-        <img src={""} alt="Minified Codewith logo"/>
+        <NavLogo src={""} alt="Minified Codewith logo"/>
         <NavItems>
-            <NavItem icon={faHome} text={"Home"} urlLocation={"/"} />
-            <NavItem icon={faCalendar} text={"Events"} urlLocation={"/events"} />
-            <NavItem icon={faBookOpen} text={"About"} urlLocation={"/about"} />
-            <NavItem icon={faBell} text={"Activities"} urlLocation={"/activities"} />
-            <NavItem icon={faTools} text={"Resources"} urlLocation={"/resources"} />
-            <NavItem icon={faUsers} text={"People"} urlLocation={"/people"} />
-            <NavItem icon={faAddressBook} text={"Contact"} urlLocation={"/contact"} />
-            <NavItem icon={faBook} text={"Code of Conduct"} urlLocation={"/code-of-conduct"} />
-            <NavItem icon={faComment} text={"Feedback"} urlLocation={"/feedback"} />
-            <NavItem icon={faQuestion} text={"FAQ"} urlLocation={"/faq"} />
-            <NavItem icon={faPalette} text={"Branding"} urlLocation={"/branding"} />
+            <NavItemLink icon={faHome} text={"Home"} urlLocation={"/"} />
+            <NavItemLink icon={faCalendar} text={"Events"} urlLocation={"/events"} />
+            <NavItemLink icon={faBookOpen} text={"About"} urlLocation={"/about"} />
+            <NavItemLink icon={faBell} text={"Activities"} urlLocation={"/activities"} />
+            <NavItemLink icon={faTools} text={"Resources"} urlLocation={"/resources"} />
+            <NavItemLink icon={faUsers} text={"People"} urlLocation={"/people"} />
+            <NavItemLink icon={faAddressBook} text={"Contact"} urlLocation={"/contact"} />
+            <NavItemLink icon={faBook} text={"Code of Conduct"} urlLocation={"/code-of-conduct"} />
+            <NavItemLink icon={faComment} text={"Feedback"} urlLocation={"/feedback"} />
+            <NavItemLink icon={faQuestion} text={"FAQ"} urlLocation={"/faq"} />
+            <NavItemLink icon={faPalette} text={"Branding"} urlLocation={"/branding"} />
             <NavItemOld>
                 <button>
                     <FontAwesomeIcon icon={faFont} color={Colours.White} />
